requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.18.0
python-dotenv==1.0.0
httpx==0.27.0
python-multipart==0.0.9
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pgvector>=0.2.0
minio>=7.2.0
fastapi-users[sqlalchemy]>=13.0.0
